+++ src/mocks/Behaviors.js
@@ -11,7 +11,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 "use strict";
 
-const Boom = require("boom");
+const Boom = require("@hapi/boom");
 
 const { compact } = require("lodash");